pyproject.toml
[tool.poetry]
name = "docker-etl"
version = "0.1.0"
description = ""
authors = ["Victor Contreras <contreras3991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pandas = "^2.2.2"
sqlalchemy = "^2.0.34"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
